Read a listing's groups fresh, and answer the review notes
- `loadGroupContext` reads the listing's groups straight from the table in
  one query instead of the shared cache. A cap edited on another edge could
  leave the cached row saying "capped" while the live capacity read said
  otherwise, and the page would have failed.
- The admin group tests are split by what they cover, so each file stays
  well under the size that keeps mutation runs narrow.
- Test fixtures use real bookings and assert a result is there before
  reading it, rather than casting or reaching through `?.`.

Fixes the type errors that failed CI: the group input now carries a real
slug index, and the attendee handler tests pass a real signed-in session.
